requirements.txt
langchain-google-genai
langchain-community
pillow
google-generativeai
tiktoken 
chromadb==0.4.4
langchain
flask
GitPython
python-dotenv
pyPDF
-e .